requirements.txt
torch>=2.2
transformers>=4.42.0
accelerate>=0.31.0
safetensors>=0.4.2
sentencepiece>=0.2.0
diffusers>=0.30.0
sentence-transformers>=3.0.0
pillow>=10.3.0,<12.0.0
soundfile>=0.12.1
pydantic>=2.7.0
numpy>=1.26
python-dotenv>=1.0.1
gradio>=5.22.0,<6.0.0
